Code restructure and clean-up
