--- index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Vulnerability Scanner</title>
-	<link rel="stylesheet" href="styles.css">
-	<style>
-		body {
-			background-image: url('background.jpg');
-			background-repeat: no-repeat;
-			background-size: cover;
-		}
-	</style>
-</head>
-<body>
-	<header>
-		<div class="container">
-			<img src="logo.png" class="logo">
-		</div>
-	</header>
-	<nav>
-		<ul>
-			<li><a href="#home">Home</a></li>
-			<li><a href="#scan">Scan</a></li>
-			<li><a href="#report">Report</a></li>
-			<li><a href="#Contact">Contact</a></li>
-			<li><a href="#logout">Logout</a></li>
-		</ul>
-	</nav>   
-	<select class="port-type">
-		<option value="select">Select Port Type:</option>
-		<option value="tcp">TCP</option>
-		<option value="udp">UDP</option>
-	</select>
-	<button class="scan-button">SCAN</button>
-</body>
-</html> 
-
