--- neterix/static/css/styles.css
@@ -1,235 +1,0 @@
-
-  
-  .text-format{
-    font-family: Arial;
-    font-size: 40px;
-    text-align: center;
-    font-weight: bold;
-    color: rgb(255, 255, 255);
-  }
-
-  
-    .scan-button {
-        background-color: #4c9baf; 
-        color: white;
-        padding: 20px 60px;
-        text-align: center;
-        text-decoration: none;
-        display: inline-block;
-        font-size: 16px;
-        margin: 0;          
-        cursor: pointer;
-        position: fixed;
-        top: 600px;
-        left: 250px;
-    }
-
-    .scan-button:active {
-      box-shadow: 0 5px #666;
-      transform: translateY(4px);
-    }
-
-         
-    
-    .scan-button:hover{
-      opacity: 0.6;
-      background-color: rgb(3, 154, 168)
-    }
-
-
-    *{
-        margin: 0;
-        padding: 0;
-        }
-        
-        
-        nav{
-          width: 100%;
-          height: 100px;
-          background-color: rgb(255, 255, 255);
-        
-        }
-        
-  
-        ul li{
-        
-          list-style: none;
-          line-height: 100px;
-          float: left;
-          position: relative;
-          left: 1000px;
-          padding-right: 20px;
-          margin-right: -5px;
-          word-spacing: 60px;
-        
-        }
-        
-        
-        ul li a {
-          display: block;
-          text-decoration: none;
-          padding:0 20px;
-          font-size: large;
-          font-family:Arial; 
-          color: rgb(19, 4, 159);
-        
-        
-        }
-
-        .port-type{
-            background-color: #4c9baf; 
-            color: white;
-            font-size: 18;
-            cursor: pointer;
-            padding: 15px;
-            position: fixed;
-            top: 500px;
-            left: 150px;
-        }
-
-        .scan-ports{
-          background-color: #4c9baf; 
-          color: white;
-          font-size: 18;
-          cursor: pointer;
-          padding: 15px;
-          position: fixed;
-          top: 500px;
-          left: 350px;}
-
-        .logo{
-            position: fixed;
-            margin-left: 10px;
-            margin-top: 10px;
-            height: auto;
-            left: 0px;
-            width:250px;
-          }
-       
-    .generate-button {
-      background-color: #4c9baf; 
-      color: white;
-      padding: 20px 60px;
-      text-align: center;
-      text-decoration: none;
-      display: inline-block;
-      font-size: 16px;
-      margin: 0;          
-      cursor: pointer;
-      position: fixed;
-      top: 500px;
-      left: 230px;
-  }
-
-  .generate-button:active {
-    box-shadow: 0 5px #666;
-    transform: translateY(4px);
-  }
-
-       
-  
-  .generate-button:hover{
-    opacity: 0.6;
-    background-color: rgb(3, 154, 168)
-  }
-
-
-  *{
-      margin: 0;
-      padding: 0;
-      }
-
-      .contact-form {
-        width: 100%;
-        padding: 12px 20px;
-        margin: 8px 0;
-        box-sizing: border-box;
-      }
-      
-      input[type=text] {
-        position: fixed;
-        font-family: verdana;
-        font-size: medium;
-        padding: 20px;
-        margin-top: 205px;
-        width: 16%;
-        margin-left: 80px;}
-      
-      input[type=email] {
-        position: fixed;
-        font-size: medium;
-        font-family: verdana;
-        padding: 20px;
-        width: 17%;
-        margin-top: 205px;
-        margin-left: 380px;}
-
-      .message-box{
-        display: block;
-        font-family: verdana;
-        font-size: medium;
-        position: fixed;
-        margin-left: 80px;
-        padding: 20px;
-        width: 36.5%;
-        height: 35%;
-        bottom: 140px;
-
-      }
-      
-      
-      .submit-button {
-        background-color: #4c9baf; 
-        color: white;
-        padding: 20px 270px;
-        text-align: center;
-        text-decoration: none;
-        display: inline-block;
-        font-size: 16px;
-        margin: 0;          
-        cursor: pointer;
-        position: fixed;
-        bottom: 60px;
-        left: 82px;
-      }
-
-      .ip-selection {
-        margin-top: 180px;
-        margin-left: 160px;
-        background: #4c9baf;
-        color: #fff;
-        border: 1px solid white;
-        padding: 25px;
-        width: 22%;
-        font-size: large;
-        text-align: center;
-        border-radius: 9px;
-        }
-
-      .ip-selection option:checked {
-        background-color: #1060d8;
-        color: #fff;}
-        
-          .ip-selection-report {
-            margin-top: 180px;
-            margin-left: 180px;
-            background: #4c9baf;
-            color: #fff;
-            border: 1px solid white;
-            padding: 25px;
-            width: 22%;
-            font-size: large;
-            text-align: center;
-            border-radius: 9px;}
-          
-            
-         .ip-selection-report option:checked {
-          background-color: #166fe3;
-          color: #fff;}
-        
-     
-        
-  
-
-
-
