--- neterix/templates/contact.html
@@ -1,42 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Vulnerability Scanner</title>
-	<link rel="stylesheet" type="text/css" href="{% static 'css/styles.css' %}">
-	<style>
-		body {
-			background-image: url('static/images/background.jpg');
-			background-repeat: no-repeat;
-			background-size: cover;
-		}
-	</style>
-</head>
-<body>
-    <p style="font-family:verdana; font-size:xx-large; color:white; position: fixed; top: 120px; left: 280px;"> Contact Us</p>
-    
-      
-      <form action="https://formsubmit.co/chesterprojectteam@gmail.com" method="POST">
-      <input type="text" name="name" placeholder="Full Name" required>
-      <input type="email" name="email" placeholder="Email Address" required>
-      <input type="hidden" name-="thanks" value ="thank you.html">
-      <textarea placeholder="Your Message" class="message-box" name="message" required></textarea>
-      <button class="submit-button">Submit</button>
-    </form>
-
-	<header>
-			<img src="static/images/logo.png" class="logo">
-		</div>
-	</header>
-    <nav>
-		<ul>
-			<li><a href="#home">Home</a></li>
-			<li><a href="{% url 'home' %}">Scan</a></li>
-			<li><a href="{% url 'report' %}">Report</a></li>
-			<li><a href="{% url 'contact' %}">Contact</a></li>
-			<li><a href="#logout">Logout</a></li>
-		</ul>
-	</nav> 
-
-
