--- neterix/templates/index.html
@@ -1,51 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Vulnerability Scanner</title>
-    <link rel="stylesheet" type="text/css" href="{% static 'css/styles.css' %}">
-	<style>
-		body {
-			background-image: url('static/images/background.jpg');
-			background-repeat: no-repeat;
-			background-size: cover;
-		}
-	</style>
-</head>
-<body>
-	<p style="font-family:verdana; font-size:xx-large; color:white; position: fixed; top: 120px; left: 175px;"> Open Port Scanner
-	</p>
-	<header>
-			<img src="static/images/logo.png" class="logo">
-		</div>
-	</header>
-	<nav>
-		<ul>
-			<li><a href="#home">Home</a></li>
-			<li><a href="{% url 'home' %}">Scan</a></li>
-			<li><a href="{% url 'report' %}">Report</a></li>
-			<li><a href="{% url 'contact' %}">Contact</a></li>
-			<li><a href="#logout">Logout</a></li>
-		</ul>
-	</nav>   
-	<select class="port-type">
-		<option value="select">Select Port Type:</option>
-		<option value="tcp">TCP</option>
-		<option value="udp">UDP</option>
-	</select>
-	<select class="scan-ports">
-		<option value="select">Select Ports:</option>
-		<option value="all_ports">All Ports</option>
-		<option value="top_thousand">Common Ports</option>
-	<select>
-	<button class="scan-button" ONCLICK="alert('Your scan has been initiated')">Scan</button>
-
-	<select class = "ip-selection" multiple>
-		{% for item in all%}
-			<option value="option1">{{item.ip_address}}</option>
-		{%endfor%}
-	  </select>
-</body>
-</html> 
-
