--- neterix/templates/report.html
@@ -1,42 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Generate Report</title>
-	<link rel="stylesheet" type="text/css" href="{% static 'css/styles.css' %}">
-	<style>
-		body {
-			background-image: url('static/images/background.jpg');
-			background-repeat: no-repeat;
-			background-size: cover;
-		}
-	</style>
-</head>
-<body>
-      <p style="font-family:verdana; font-size:xx-large; color:white; position: fixed; top: 120px; left: 150px;"> Generate Port Activity Report
-      </p>
-    </div>
-    <header>
-			<img src="static/images/logo.png" class="logo">
-		</div>
-	</header>
-	<nav>
-		<ul>
-			<li><a href="#home">Home</a></li>
-			<li><a href="{% url 'home' %}">Scan</a></li>
-			<li><a href="{% url 'report' %}">Report</a></li>
-			<li><a href="{% url 'contact' %}">Contact</a></li>
-			<li><a href="#logout">Logout</a></li>
-		</ul>
-	</nav>   
-    <button class="generate-button" ONCLICK="alert('Your report has been generated and emailed to the administrator')"
-    >Generate Report
-    </button>
-	<select class = "ip-selection-report" multiple>
-		{% for item in all%}
-			<option value="option1">{{item.ip_address}}</option>
-		{%endfor%}
-	  </select>
-	<body>
-	<html>
