--- neterix/urls.py
@@ -1,11 +1,0 @@
-
-from django.urls import path
-from .import views
-
-urlpatterns = [
-   path('', views.home, name="home"),
-   path('contact', views.contact, name="contact"),
-   path('report', views.report, name="report"),
-]
-
-
