--- styles.css
@@ -1,94 +1,0 @@
-
-  
-  .text-format{
-    font-family: Arial;
-    font-size: 40px;
-    text-align: center;
-    font-weight: bold;
-    color: rgb(255, 255, 255);
-  }
-
-  
-    .scan-button {
-        background-color: #4c9baf; 
-        color: white;
-        padding: 20px 40px;
-        text-align: center;
-        text-decoration: none;
-        display: inline-block;
-        font-size: 16px;
-        margin: 0;          
-        cursor: pointer;
-        position: fixed;
-        top: 600px;
-        left: 160px;
-    }
-
-         
-    
-    .scan-button:hover{
-      opacity: 0.6;
-      background-color: rgb(3, 154, 168)
-    }
-
-
-    *{
-        margin: 0;
-        padding: 0;
-        }
-        
-        
-        nav{
-          width: 100%;
-          height: 100px;
-          background-color: rgb(255, 255, 255);
-        
-        }
-        
-  
-        ul li{
-        
-          list-style: none;
-          line-height: 100px;
-          float: left;
-          position: relative;
-          left: 1000px;
-          padding-right: 20px;
-          margin-right: -5px;
-          word-spacing: 60px;
-        
-        }
-        
-        
-        ul li a {
-          display: block;
-          text-decoration: none;
-          padding:0 20px;
-          font-size: large;
-          font-family:Arial; 
-          color: rgb(19, 4, 159);
-        
-        
-        }
-
-        .port-type{
-            background-color: #4c9baf; 
-            color: white;
-            font-size: 18;
-            cursor: pointer;
-            padding: 15px;
-            position: fixed;
-            top: 500px;
-            left: 150px;
-        }
-
-        .logo{
-            position: fixed;
-            margin-left: 10px;
-            margin-top: 10px;
-            height: auto;
-            left: 0px;
-            width:250px;}
-  
-
-
