+++ website/Neterix/urls.py
@@ -1,0 +1,10 @@
+
+from django.contrib import admin
+from django.urls import path, include
+
+
+urlpatterns = [
+    path('admin/', admin.site.urls),
+    path('', include('website.Scanner.urls')),
+    path('accounts/', include('django.contrib.auth.urls'), name="accounts")
+]
