+++ website/Scanner/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Cve, Ip_cve, Ip_port, Scan, Ip, Port
+
+
+admin.site.register([Cve, Ip_cve, Ip_port, Scan, Ip, Port])
