+++ website/templates/main/home.html
@@ -1,0 +1,8 @@
+{% extends 'base.htm' %}
+
+{% block main %}
+<div class="container text-dark">
+    <h1>Home</h1>
+    <p>Home page</p>
+</div>
+{% endblock %}
