+++ website/templates/main/report.html
@@ -1,0 +1,14 @@
+{% extends 'base.htm' %}
+{% load static %}
+
+
+{% block main %}
+<button class="generate-button"
+    ONCLICK="alert('Your report has been generated and emailed to the administrator')">Generate Report
+</button>
+<select class="ip-selection-report" multiple>
+    {% for item in all%}
+    <option value="option1">{{item.ip_address}}</option>
+    {%endfor%}
+</select>
+{% endblock %}
