+++ website/templates/main/scan.html
@@ -1,0 +1,21 @@
+{% extends 'base.htm' %}
+{% load static %}
+
+{% block main %}
+<select class="port-type">
+    <option value="select">Select Port Type:</option>
+    <option value="tcp">TCP</option>
+    <option value="udp">UDP</option>
+</select>
+<select class="scan-ports">
+    <option value="select">Select Ports:</option>
+    <option value="all_ports">All Ports</option>
+    <option value="top_thousand">Common Ports</option>
+    <button class="scan-button" ONCLICK="alert('Your scan has been initiated')">Scan</button>
+</select>
+<select class="ip-selection" multiple>
+    {% for item in all%}
+    <option value="{{item.ip_address}}">{{item.ip_address}}</option>
+    {%endfor%}
+</select>
+{% endblock %}
