+++ website/templates/registration/login.html
@@ -1,0 +1,10 @@
+{% extends 'base.htm' %}
+
+{% block main %}
+  <h2>Login</h2>
+  <form method="post">
+    {% csrf_token %}
+    {{ form.as_p }}
+    <button type="submit">Login</button>
+  </form>
+{% endblock %}
